Done Asset API

+++ src/main/java/lk/ecommerce/zeetradexbackend/service/impl/OrderServiceImpl.java
@@ -5,6 +5,7 @@ import lk.ecommerce.zeetradexbackend.enums.OrderStatus;
 import lk.ecommerce.zeetradexbackend.enums.OrderType;
 import lk.ecommerce.zeetradexbackend.repo.OrderItemRepo;
 import lk.ecommerce.zeetradexbackend.repo.OrderRepo;
+import lk.ecommerce.zeetradexbackend.service.AssetService;
 import lk.ecommerce.zeetradexbackend.service.OrderService;
 import lk.ecommerce.zeetradexbackend.service.WalletService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -27,6 +28,9 @@ public class OrderServiceImpl implements OrderService {
 
     @Autowired
     private OrderItemRepo orderItemRepo;
+
+    @Autowired
+    private AssetService assetService;
 
     @Override
     public Order createOrder(User user, OrderItem orderItem, OrderType orderType) {
@@ -85,6 +89,15 @@ public class OrderServiceImpl implements OrderService {
          Order savedOrder = orderRepo.save(order);
 
          //Creating asset
+        Asset oldAsset = assetService.findAssetByUserIdAndCoinId(
+                order.getUser().getId(), order.getOrderItem().getCoin().getId());
+
+        if (oldAsset == null) {
+            assetService.createAsset(user, orderItem.getCoin(), orderItem.getQuantity());
+
+        }else {
+            assetService.updateAsset(oldAsset.getId(), quantity);
+        }
 
         return savedOrder;
     }
@@ -98,27 +111,42 @@ public class OrderServiceImpl implements OrderService {
         }
         double sellPrice = coin.getCurrentPrice();
 
-        double buyPrice = assetToSell.getPrice();
+        Asset assetToSell = assetService.findAssetByUserIdAndCoinId(user.getId(), coin.getId());
 
-        OrderItem orderItem = createOrderItem(coin, quantity, buyPrice, sellPrice);
+        double buyPrice = assetToSell.getBuyPrice();
 
-        Order order = createOrder(user, orderItem, OrderType.SELL);
-        orderItem.setOrder(order);
+        if (assetToSell != null) {
 
-        if (assetToSell.getQuantity()>= quantity) {
-            walletService.payOrderPayment(order, user);
 
-            order.setStatus(OrderStatus.SUCCESS);
-            order.setOrderType(OrderType.SELL);
-            Order savedOrder = orderRepo.save(order);
+            OrderItem orderItem = createOrderItem(coin,
+                    quantity,
+                    buyPrice,
+                    sellPrice);
 
-            Asset updatedAsset = assetService.updateAsset(assetToSell.getId(), -quantity);
-            if (updatedAsset.getQuantity()*coin.getCurrentPrice() <=1) {
-                assetService.deleteAsset(updatedAsset.getId());
+
+
+            Order order = createOrder(user, orderItem, OrderType.SELL);
+            orderItem.setOrder(order);
+
+            if (assetToSell.getQuantity() >= quantity) {
+
+                order.setStatus(OrderStatus.SUCCESS);
+                order.setOrderType(OrderType.SELL);
+                Order savedOrder = orderRepo.save(order);
+
+                walletService.payOrderPayment(order, user);
+
+                Asset updatedAsset = assetService.updateAsset(assetToSell.getId(), -quantity);
+
+
+                if (updatedAsset.getQuantity() * coin.getCurrentPrice() <= 1) {
+                    assetService.deleteAsset(updatedAsset.getId());
+                }
+                return savedOrder;
             }
-            return savedOrder;
+            throw new Exception("Insufficient quantity to sell");
         }
-        throw  new Exception("Insufficient quantity to sell");
+        throw new Exception("asset not found");
 
 
     }
